pyproject.toml
[project]
name = "raindropper"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
raindropper = "raindropper.main:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
